02-DApp/app/requirements.txt
Flask
pandas
web3
python-dotenv
